requirements.txt
Flask
flask-socketio
pycryptodome
requests
python-socketio
